requirements.txt
os
requests
streamlit
torch
selenium
selenium-stealth
webdriver-manager
pillow
